feat: type the SMP sequence as u8 (#70)
`Data.to_frame()` and `SMPRequest.to_frame()` now take `sequence: u8`
rather than `int`, so the header's 8 bit sequence field is enforced by
the type checker instead of by a `struct.error` at pack time:

    request.to_frame(0x100)

    mypy:    error: Argument 1 to "to_frame" has incompatible type
             "Literal[256]"; expected "Literal[0, 1, ..., 255]"  [arg-type]
    pyright: error: Argument of type "Literal[256]" cannot be assigned to
             parameter "sequence"  (reportArgumentType)

Dropping the module-global counter moved the `% 0x100` wrap from the
library to the caller. `u8` is what makes that transfer safe: the bound
the counter used to apply silently is now part of the signature callers
program against.

types-bits
----------
`types_bits.u8` is a fully materialized `Literal` union shipped as a
`.pyi`, so this is a type-checker-only construct with no runtime cost --
the import sits under `TYPE_CHECKING` in both modules and never executes.

It is a `[project]` dependency rather than a dev dependency because `u8`
appears in `smp`'s public signatures and `smp` ships `py.typed`, so
downstream checkers must be able to resolve it.

0.2.0 is the first release usable here. 0.1.0 was blocked twice over: it
declared `Requires-Python: >=3.12` against smp's `>=3.10`, and its stub
used PEP 695 `type u8 = ...`, which mypy rejects outright whenever
`--python-version` is below 3.12 --

    types_bits/__init__.pyi:8: error: Type statement is only supported in
    Python 3.12 and greater  [syntax]

-- so the 3.10 and 3.11 matrix legs would have failed even with the floor
lowered. 0.2.0 lowers the floor and emits `u8: TypeAlias = Literal[...]`.

warn_unused_ignores
-------------------
Enabled for mypy so the annotation is load-bearing rather than
decorative. `test_to_frame_rejects_out_of_range_sequence` covers the
unchecked caller by passing -1 and 0x100 behind a
`type: ignore[arg-type]`; with this flag, reverting `sequence: u8` to
`int` makes that ignore unused and fails the typecheck task:

    tests/test_message.py:142: error: Unused "type: ignore" comment  [unused-ignore]

The repo was already clean under the flag otherwise.

Wire format is unchanged; `camas matrix` is green on Python 3.10-3.15.

+++ tests/test_generics_typing.py
@@ -29,6 +29,7 @@ from smp import zephyr_management as zephyr
 from smp.user import intercreate as ic
 
 if TYPE_CHECKING:
+    from types_bits import u8
     from typing_extensions import TypeIs
 
     from smp import error as smperror
@@ -64,7 +65,7 @@ def _request(request: SMPRequest[TRep, TEr1, TEr2]) -> TRep | TEr1 | TEr2:
 def _client_owned_frame(
     request: SMPRequest[TRep, TEr1, TEr2],
     *,
-    sequence: int,
+    sequence: u8,
     version: smpheader.Version,
     flags: smpheader.Flag,
 ) -> smpmessage.Frame[Any]:
